Update new card
+now "TOUGH BODY " card is available (ID: 3)
+change all damage to float type instead of int
+put UIDisplay after default time(for health update)

+++ Assets/Scripts/DamageDealer.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class DamageDealer : MonoBehaviour
 {
-    [SerializeField] int damage = 10;
+    [SerializeField] float damage = 10;
 
-    public int GetDamage()
+    public float GetDamage()
     {
         return damage;
     }
